requirements.txt
horovod
numpy
pytorch
tensorboard
tqdm
